Move component to seperate category
